Ensure crate IDs are correct
Summary:
rust-project.json uses indexes in when listing dependencies, e.g.:

```
"crates": [
  {
    "display_name": "my-project",
    "deps": [
      {
        "crate": 1,
        "name": "my-lib"
      }
    ]
  },
  {
    "display_name": "my-lib",
    "deps": []
  }
]
```

This means we must iterate through the crates/targets in a consistent order, or the indexes will be wrong. Previously, we iterated in `FxHashMap` order, which did not maintain this invariant (it was a `BTreeMap` prior to D55881958).

Instead, build a `Vec<Target>` inside `to_json_project` and use the array indices. Remove

Reviewed By: davidbarsky

Differential Revision: D56084301

fbshipit-source-id: 0e28b3e8d60222e397f2e0039079afa4e8b3c34d