Revert D56138785: Multisect successfully blamed "D56138785: [go][buck2] Use symlinks in `go list`" for one test failure
Summary:
This diff reverts D56138785
D56138785: [go][buck2] Use symlinks in `go list` by podtserkovskiy causes the following test failure:

Tests affected:
- [cogwheel:cogwheel_fbdns_test#main](https://www.internalfb.com/intern/test/844424952761408/)

Here's the Multisect link:
https://www.internalfb.com/multisect/4867143
Here are the tasks that are relevant to this breakage:

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Reviewed By: podtserkovskiy

Differential Revision: D56177780

fbshipit-source-id: a6a0e0c0f989f359cc24be8dc54e557617da484f