Make shared lib soname opaque (take 2)
Summary:
Use a new, opaque type for the`soname` field on `SharedLibrary`,
which can hold either a static `str` or an `Artifact` that contains a
build-time generated SONAME, and updated users that require a
static SONAME to enumerate this clearly in the code.

This will allow migrating some rules to support SONAMEs which aren't
known at analysis time (see D55000520).

Reviewed By: dtolnay

Differential Revision: D55369940

fbshipit-source-id: eaa97bdf4483af1d97634807429f2783e534a07e