interpreter: Don't make universal globals configurable
Summary: I don't feel super strongly about this being an improvement. That being said, I do think that it's *generally* good if we don't make things configurable only for tests. Also, there's a significant amount of extra code that just has to deal with passing this thing around that I'll clean up in the next few diffs, so it seems like an improvement to not have all that

Reviewed By: stepancheg

Differential Revision: D56225330

fbshipit-source-id: ba46a384bf30d5a3d2d4c3ad39bc2596b9b9ac97