Revert D56029241: Multisect successfully blamed "D56029241: [buck2][PR] Use transitive set for HaskellLibraryInfo" for one test failure
Summary:
This diff reverts D56029241
D56029241: [buck2][PR] Use transitive set for HaskellLibraryInfo by ndmitchell causes the following test failure:

Tests affected:
- [cogwheel:cogwheel_sigma_tailer_regression_test#main](https://www.internalfb.com/intern/test/844424972293552/)

Here's the Multisect link:
https://www.internalfb.com/multisect/4878252
Here are the tasks that are relevant to this breakage:

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Reviewed By: scottcao

Differential Revision: D56227800

fbshipit-source-id: 0cc8d511ab34e8954cdbc8e37faef9e73034c859