errors: Rename `User` -> `Input`, `Infra` -> `Tier0`
Summary:
Again, as per the RFC.

I'm not 100% certain on the naming here. For tier1 errors, we have a clearly agreed upon name in "bad experiences." We never really agreed on such a name for tier0 errors or tier2 errors. For tier2 we at least have "input" as a working suggestion, while for tier0 we have nothing. Therefore my current plan is to have it as "input," "bad experience," and "tier0" in code.

I have yet to decide what to do about telemetry.

Reviewed By: stepancheg

Differential Revision: D56229713

fbshipit-source-id: b2b7cef0c90f9a69ae2c9990bfd8f216cf08c3b6