cli: Remove two unused flags
Summary:
These flags only have an effect if the user is also using `--no-buckd` or explicitly invoking `buck2 daemon`, which really doesn't seem terribly useful. Just make it explicit that this can only be configured via an env var

The `--help` changes from this diff are a bit surprising. Not going to worry too much about it now, it's not a huge regression and we should use headings to prevent it properly

Reviewed By: stepancheg

Differential Revision: D56348238

fbshipit-source-id: 5e393413d23a22c6e080f0b331956af4718546fe