cli: Move `DaemonBeforeSubcommandOptions` to daemon command
Summary:
The only thing that we have to do to make this work is pick a default value for the `--no-buckd` case. But these flags are only used to reject certain daemons, and so they can't really be expected to do anything for `--no-buckd` anyway.

Next diff fixes some names and cleans up

Reviewed By: stepancheg

Differential Revision: D56348236

fbshipit-source-id: 73fbde8807ea75b7d405fae4f2c9471707775e65