env: Introduce `Applicability`
Summary:
I want to be able to mark env vars by whether they're only for testing or not for OSS users. So introduce a value for that, and add support for parsing it from the macro. Next diff will actually use it.

The macro shenanigans needed to support this are a little bit complicated, but not excessively so. They're also fairly composable, so won't get more complicated if we want to add additional stuff in the future

Reviewed By: stepancheg

Differential Revision: D56391374

fbshipit-source-id: cf6cbe8f9d2cdc20030e6ac742e151e334a26895