handle regular file in subpath for `fs_util::remove_all`
Summary:
Handle case when path passed to `remove_all` contains a subpath with a
regular file.

Before this diff, if you `remove_all("/foo/bar")` when `/foo` is a regular
file, you get `io::ErrorKind::NotADirectory` error. After this diff, it's a nop
returning `Ok`, which seems natural and should not lead to any behavior changes
as original path is non-existent indeed.

Reviewed By: JakobDegen

Differential Revision: D56389601

fbshipit-source-id: a514b1634bceae7b7c215efcc047877547ec933a