cli: Add some option headings
Summary:
For a selection of our command line options, add headings to them in the help output. This makes it easier for users to find related options.

The `UnsetNextHelpHeading` thing is pretty ugly but unfortunately it's the only real way to do this in clap.

Note that this diff does come with some regressions in argument ordering, specifically all the positional arguments are now at the bottom. That's bad, but I'm not going to worry about it right now. Later in this stack I'm redoing all the argument ordering anyway, and this will be fixed as a part of that

Reviewed By: stepancheg

Differential Revision: D56446630

fbshipit-source-id: 7cc52820b6883cc3386cc5c3ff22a16e8d2d08d5