cli: Reorder help sections
Summary:
This diff goes through all our command definitions and reorders fields so that the least interesting argument groups (primarily, those that apply to every command) are at the end. This change is reflected in the help output - those uninteresting argument groups are now reliably at the bottom of the `--help`, with the command specific options at the top. This seems like a strong improvement to me

This diff also fixes the regression re positional args not appearing at the top

Reviewed By: stepancheg

Differential Revision: D56446632

fbshipit-source-id: 58bb1558ddbd1023e6fe5ffeb555d399ddadd977