cli: Hide some useless options
Summary:
The `--skip-duplicate-targets` option isn't supposed to exist anyway, so don't tell users about it. One day we'll be able to delete it.

This also hides the options for unused target cfgs. We support specifying those on `uquery` and friends for modefile reasons, but they don't do anything so don't show them in `--help`

Reviewed By: lmvasquezg

Differential Revision: D56446629

fbshipit-source-id: cec07e1de0e7665f1982d897de9f84b6fbd191d9