Fix client id missing in scuba log for client commands
Summary:
Any command running locally was failing to forward along the argument

--client-metadata id=some_id

Added a unit test (test_client_metadata_clean) to verify the metadata is properly appended for a clean command.

In order to run the unit test properly, I had to make it so clean, kill, and killall are able to write their metadata to a file, which would again only work on server based commands before this diff.

Updated help tests for kill and killall to include the new event log options for testing the metadata.json.

Reviewed By: JakobDegen

Differential Revision: D56653040

fbshipit-source-id: 985624195572d309eb2422972b1e144ef3cd575b