Rename the absolute argsfile to xcode argsfile
Summary:
When we created the absolute argsfile in the D46334207 stack, it was to be used in Xcode, but we attempted to keep the logic generic. However, since then, the argsfile has
1. only been used in Xcode
2. added Xcode specific changes

Further up the stack, we'll be changing the argsfile to infact revert to using relative paths, but with Xcode specific changes. As such, rename it.

Extract the rename into it's own diff to handle updating relevant tests and flows that consume it.

It would be nice to update the Cxx layer to perhaps provide lamda functions to generate additional argsfiles/content, etc and make it more plug and play vs directly integrate. But that doesn't seem to get us anything at this moment, without seeing other needs to make it generic.

Reviewed By: rmaz

Differential Revision: D56731573

fbshipit-source-id: 7dedb5f441a188e68b65205b4861a59325617ce6