support symlinks to non-files in --out
Summary:
as per title, `fs::copy` doesn't support copying symlinks which are not targeting regular files. Add a special handling for symlinks. Just recreate symlinks in a destination directory and overwrite anything pre-existing there (keeping the same behavior as `fs::copy` has).

If one of the output artifacts contains a symbolic link pointing outside of this artifact, convert it to absolute link to avoid broken symlinks in `--out` directory when their counterparts in `buck-out` are valid ones.

Reviewed By: JakobDegen

Differential Revision: D56184904

fbshipit-source-id: be9f03b87b9e908c7320f2a9e87a63cf2962863d