external cells: Declare source files to materializer
Summary:
Like it says in the title.

The implementation strategy I went for here is to do this eagerly the first time the external cell is accessed. This is convenient because it matches what will happen with git/http cells, as those will need to be downloaded at that point.

Reviewed By: stepancheg

Differential Revision: D55792067

fbshipit-source-id: b8a1bb813d124ec6e51e91a9f2ed1e5817a579e5