external cells: Set up bundling for nano-prelude
Summary:
In addition to the standard prelude, the nano prelude is a good candidate for a bundled cell. It results in both the advantages that 1) it's a bit easier to set up and 2) you can `cd` into the data directory and invoke `buck` commands directly to debug tests.

This diff implements most of the logic to actually do the bundling in buck-powered builds. The implementation turns out to be slightly complicated because rustc insists on interpreting `include!`s relative to the source file doing the including.

Cargo based builds are not yet included in this diff, and will be added later. I'll do a separate buildscript-powered implementation for those, which will be a lot simpler on account of 1) caring less about correctness, 2) being able to use absolute paths, and 3) not planning to add support for compression, since I don't want to have to support cargo in the long term.

Reviewed By: stepancheg

Differential Revision: D55792077

fbshipit-source-id: 19cd3b16aa1b89b6b570fb520f0409c68e26c292