external cells: Resolve source paths for external cell sources into buck_out
Summary:
In order to support using source files from external cells for running actual actions, we'll need to be able to actually write the files to disk somewhere. This diff does that, by modifying the resolution of `PackagePath` and friends to point to `buck-out/v2/external/bundled/bundled_cell_name/package/src.txt`.

Of course, this diff does not yet actually materialize those files to disk - we'll do that next.

The approach taken in this diff is fairly robust in that it covers all the resolution of `PackagePath` to `ProjectRelativePathBuf`. The `CellResolver` still has some methods for resolving `CellPath`s to `ProjectRelativePathBuf`s. In principle, these methods might be wrong with external cells, since the source files aren't actually located in the repo on disk. However, I looked through all the call sites and they all actually do need the fake in-repo path, and not the buck-out path. So that seems pretty good.

Reviewed By: stepancheg

Differential Revision: D55792083

fbshipit-source-id: cf0aea53c4d80e05962fde61f784cce53a587c3b