Create fake action for uploaded dep file, use digest as remote dep file key
Summary:
A security check was added by alessandrolipa to the action cache last year that causes the action cache to fetch the original action to check the platform field whenever an action result is uploaded.
Currently when we upload a remote dep file, we re-upload the original action, then upload a new action result for that action but with the dep file key as the digest. In reality the dep file key is not the digest for the original action so the action look up fails.

For dep file action results, the dep file is not actually used, so this constructs a fake action so that the action digest/dep file key is a digest of the original dep file key and the platform.
This means we can't get remote dep file cache hits from different RE platforms but I assume that is already the case for normal action cache hits.

Reviewed By: scottcao

Differential Revision: D56449722

fbshipit-source-id: f89ad560067d12d5755228d2685816ee10316d21