Reuse action result for dep file upload instead of uploading twice
Summary:
We should not try to re-upload the outputs of actions when uploading remote dep file metadata.
For local actions this isn't broken but it's unnecessary and likely slow.
For remote actions it doesn't work because outputs are not materialized at this point and shouldn't be, and we need a way to re-use an existing action result (added later).

Splitting `perform_cache_upload` into `upload_local_outputs` and `upload_dep_files` to support this, they share some code but at this point they are different enough that it's more confusing to use a single method.

The main differences are:
- `upload_local_outputs` actually uploads action outputs and returns the action result for uploaded outputs.
- `upload_dep_files` appends metadata to an existing action result and uploads it, instead of creating a new `TActionResult2` by re-uploading outputs. It also skips the max size check since no outputs are uploaded.

Reviewed By: scottcao

Differential Revision: D56289406

fbshipit-source-id: 9ac06ef38a056c0c510713300d96a83f0da4811e