materializer: Add `has_artifact_at`
Summary:
The context for this diff is that I want to be able to avoid downloading git-backed external cells on every single daemon restart. In order to achieve that I need to be able to, on a daemon restart, find the contents of the external cell on disk somewhere and ensure that they're correct.

The basic plan for doing that is to put the commit hash of the external repo into the file path. After downloading, I can then call `declare_existing` and that should mostly work; unfortunately though, that doesn't get us quite all the way there, because clean stale might in general delete parts of that directory out from under me, and I'll have no way to tell that that happened.

This diff will let me use clean stale's invalidation of the materializer state to prevent that.

Reviewed By: christolliday

Differential Revision: D56148812

fbshipit-source-id: 1b4e9c8210204c4fd47e7c712ba0022283e49990