external cells: Repro bug
Summary:
https://github.com/dtolnay/cxx/pull/1345#issuecomment-2101961930

This is kind of a bad test. The reason that this was able to slip through in the first place is because most of this stuff is tested in integration tests, which aren't run on cargo builds of buck2. But writing a better test is not super trivial, and I'd rather spend time deprecating cargo builds

Reviewed By: samkevich

Differential Revision: D57148214

fbshipit-source-id: 38e5148a2c0b2436dab6ab2a70cd5dbc5b01d185