buck2 | xcframeworks | Add module_map as a field to apple_bundle
Summary:
Allows adding custom module_map to [xc]frameworks.

We probably want to also allow using an autogenerated module.modulemap file too, but this unblocks work on WA*Kit frameworks, and we potentially want to allow customization here anyway.

Should the field be `module_map` or `modulemap`?
The extension of the file is indeed `modulemap`, but the type of the file is 'module map', so I'm using the underscore.

Reviewed By: chatura-atapattu

Differential Revision: D57209531

fbshipit-source-id: 642749bbba36563a4fa5d4e3cb90227767053709