Remove effectively unused is_shared parameter
Summary:
After D41708290, the is_shared parameter has no effect. That is while code branches on it's value, the possible branches are identical.
We see that based on the value of `is_shared` a shared variant of projections may be called (ie. `default` vs `default_shared`, `stripped` vs
`stripped_shared`). Those shared variants are identical to the original.

This is an NFC that simplifies the code in preparation for the stacked changes.

Reviewed By: rmaz

Differential Revision: D57050799

fbshipit-source-id: 732aa37df4a233fa7f67a1514bd3eb4611a8adf0