Actually track parallel deps
Summary:
This changes the ModernComputeCtx compute functions actually produce
contexts that "correctly" track parallel deps (they still are flattened into a
list, but that'll be updated soon).

The approach depends on typed-arena, which basically allows us to have a
Vec<RecordedDeps> where each parallel context can reserve a slot to write its
deps into, and then the first time that the outer context interacts with it after that
it means the parallel bits are done and it can see all the parallel deps in the
arena.

This is kinda undoing D53678531 (which deleted the initial parallel
implementation), but the implementation is actually quite different.

This is the first step in getting the recomputation to respect
the inter-dep data flow dependencies. D55547362 adds tests
for the two bits of behavior that this affects:
1. in a series of deps, we shouldn't compute past the first changed dep
2. we can compute through to the end of parallel series when one of them has a change

And then D55547388 changes the recomputation to have that
"correct" behavior

Reviewed By: stepancheg

Differential Revision: D55547381

fbshipit-source-id: 858be74eb33b0ee4c9ae610168d16739366e08f9