Add SeriesParallelDeps
Summary:
This changes RecordingDepsTracker to store deps as a SeriesParallelDeps which
actually preserves the inter-dep data flow.

Includes an efficient encoding of the SeriesParallelDeps as a pair of
`(Vec<DiceKey>, Vec<u32>)` where the second vec encodes the structure of the
graph.

There's an SPEncoder/SPDecoder for that encoding.

There's iterators that show how we can iterate over that graph structure.

From the docs:

The DiceComputations compute apis are designed so that in normal usage the graph of
inter-dep data dependencies within a compute node form a series-parallel graph.

The [SeriesParallelDeps] records the deps and the structure of that graph so that
when we recompute we can check keys in parallel but avoid requesting a key that
would not be requested by calling the compute directly in that state.

For non-trivial graphs, we will encode the graph as a flat list of keys and an
encoding of the description of the graph.

The [SeriesNodeIterator] and [ParallelNodeIterator] provide a fairly readable
implementation of decoding and traversing the graph.

A Series node is a sequence of keys and parallel nodes. It's encoding is a sequence of two items:
  SPItem::Key(v): Indicates that the next v keys are part of this series
  SPItem::Parallel{keys: x, specs: y}: Indicates that the next `x` keys and `y` specs form a parallel node

If a Series node has remaining keys not covered by its spec, the remaining keys are all part of the series.

A Parallel node is a set of Series nodes. A parallel node of series nodes S1, S2, and S3 will have:

```ignore
keys = keys(S1) + keys(S2) + keys(S3)
spec = header(S1) + spec(S1) + header(S2) + spec(S2) + header(S3) + spec(S3)
```

Where `spec(S1)` is the encoding of a series Node as described above and `header(S1)` is a SPSeriesHeader, which
is one of two items:
  SPSeriesHeader::Simple(n): Indicates the series has no nested parallel nodes and is just a series of n keys (and 0 specs).
  SPSeriesHeader::Complex{keys: x, specs: y}: Indicates a complex series that covers the next x keys and y specs.

For both SPItem::Parallel and SPSeriesHeader::Complex, the specs value is the size of the encoded specs.

Reviewed By: stepancheg

Differential Revision: D55547376

fbshipit-source-id: e50856c55a65129edfe6bac5de4bc8462bed9bfe