Explicitly define fbsource and waios build modes
Summary:
Previously, we were attempting to be dynamic in the build modes defined for FBSource/WAIPhone/OSS. However, in the current world of configuration (buck config, constraints, target platforms and config modifiers), this became awfully complex to get right in every situation.

Especially when attempting to define selects that are shared in different repositories (simple example in the next diff).

Thus, lets explicitly define the build modes used in fbsource and whatsapp. The open source ones were deleted in D55526692.

FBSource:
- local
- master
- profile
- rc
- production

Whatsapp:
- debug
- profile
- beta
- release

Reviewed By: rmaz

Differential Revision: D57397477

fbshipit-source-id: 10099c92aa95924321ae20bf5e6489f24475bb8e