Measure and log install_duration to buck2_builds
Summary:
`install_duration` is the measured as the time between the last action is executed and materialized, and the last send_file event is returned from the installer as installation time, that is, the time saved if installation was instant.

Ultimately this should be added to the critical path data set, but at that point we would still want to aggregate the install events in the critical path in some way, record it to buck2_builds, and validate that the aggregation (roughly) corresponds to this duration, so this seems worth adding regardless.

Specifically, this should be equivalent to the total 'potential improvement time' for every install event in the critical path (not the total execution time), if there was a critical path event for each `send_file` + `send_install_info` event.

Reviewed By: IanChilds

Differential Revision: D57362423

fbshipit-source-id: f6f793861d4c16ce7ba93864b56f24242abf050e