Start to consolidate logic under _rust_providers and _native_providers
Summary:
`_native_providers` has a dependency on the behavior in `_rust_link_providers`. In particular, the former makes assumptions about the latter, which would be clearer / simpler to maintain if their logic were collocated.

For example:
* "The rust link providers already contain the linkables for the `archive` and `pic_archive` cases"
* "The node that we create here cannot just use this target's label, since that would conflict with the node created for the native providers."

Especially when building with `advanced_unstable_linking`, these functions have fairly high overlap, so I intend to pull out that common logic and pass it to `_rust_link_providers` / `_native_link_providers` as needed.

Reviewed By: dtolnay

Differential Revision: D57417295

fbshipit-source-id: 82391ec09cbb405c9a24752d9a7fe5eee366ce26