buck2: explain: wrap everything under Router
Summary:
Realized that I can't rely on <a> links, because page reloads are [slow when serving from Manifold](https://fb.workplace.com/groups/812395560360805/permalink/815782080022153/).

The idea is to abstract url management. In order for that logic to live in the Router, I need to put all the components that may use under it

Reviewed By: samkevich

Differential Revision: D57424335

fbshipit-source-id: 52fc92bf5c034bc18ee2d7608e138c9915ab3045