buck2 | xcframework | Optionally copy -Swift.h
Summary:
## What this diff does:
It copies any generated -Swift.h header for the single dependency that is given as the `binary` field in the `apple_bundle` for a framework iff `copy_public_framework_headers` is set to `True`.

## What this diff does not do:
Generate and include a -Swift.h header that reflects all public/open Swift entities in the full dependency graph.
1) This seems hard to do
2) This might be counter-productive as the pattern for Swift+Buck seems to be to make each buck target correspond to a module. With this pattern the right thing to do would be to create a separate [xc]framework per module.

Reviewed By: milend

Differential Revision: D57215035

fbshipit-source-id: 2e4823cf75e182c8ddabf1a87c87aaacb2f94d83