remove bad with_context
Summary:
This was intended to be context on an error returned from checking
is_ignored, but was accidentally actually context on the error from converting
the is_ignored result to an error (i.e. is_ignored returned
anyhow::Result<IgnoreResult>, and IgnoreResult in turn converts to a
Result). Now the outer Result has actually been removed, but this context is
still here.

Anyway, it should go away.

Reviewed By: JakobDegen

Differential Revision: D56343500

fbshipit-source-id: 06389fa2a2b436bd2f39ccc7aa4aed2d8262fa95