Create and enable darwin version of dist_lto_tools for apple_toolchain
Summary:
Distributed thin-lto uses a series of wrapper sripts to preform different steps
in the build process. The requirements for Darwin are different, so we introduce
and create a different set of tools for Darwin, and enable it in apple_toolchain.

The diff on it's own should be an nfc as no apple_toolchain powered build should use distributed thin-lto at the moment.

Reviewed By: rmaz

Differential Revision: D56939412

fbshipit-source-id: f5c322c349edd35388f2fda08a20d9f130620624