Add and use non-filelist LinkInfo TSet projections for Darwin linker invocations
Summary:
Until now, our Darwin linker invocations have consisted of linker flags and a filelist. Certain "linkables" can be listed as part of a filelist, or directly on the command line. In particular, object files and static archives may appear in either.

When a Darwin linker invocation is crafted, we need to convert a set of linkables into command line arguments. To do this, we project a TSet of LinkInfos twice, once for filelist arguments, once for direct command line arguments.

If linkables that can appear in both places are included in both projections, we will include them twice in the linker invocation.

The solution to this has been to branch in `append_linkable_args` if we are using a Darwin linker such that we don't include these linkables directly in the command line, only in the filelist.

The problem with this is that we now have cases (for distributed thin-lto) where we want to craft a Darwin linker invocation without using a filelist. We want to be able to re-use the same APIs to craft non-filelist based Darwin linker invocations.

This diff adds more projections that allows the user to request all "non filelist" arguments explicitly. This way when crafting a linker invocation, we have the option to request "all", "filelist" or "non-filelist" arguments and combine them as we see fit. When crafting the linker invocation, we request either all arguments, or filelist and non-filelist arguments depending on if we are using a filelist or not.

This diff attempts to add the new functionality without disturbing existing code, but creates some duplication. Future diffs will refactor this into common code, but this way is hopefully easier to review and test.

Reviewed By: rmaz

Differential Revision: D57050796

fbshipit-source-id: 1038586dceff2089a2f25120b174460a53f5de0a