fix flaky installer duration test
Summary:
This test is flaky because `time_to_last_action_execution_end_ms` records the time when the invocation recorder receives the last action execution event, not when the event happens, which can be too late since install time + time to last action execution is already very close to the total command duration.

The way buck currently measures `time_to_last_action_execution_end_ms` is slightly cheaper than doing it based on timestamps and happens for every action, so fixing this in the test instead of in buck.

Reviewed By: iguridi

Differential Revision: D57505903

fbshipit-source-id: dea0932bbdeaeabdb1e9b42e954e94254240fcae