Update tests to not assume a particular behavior
Summary:
A lot of tests assume that we treat having a list of nodeps different
than having a list of one dep that never changes. Ideally, if we compute a node
that has no deps, we should be able to reuse that value at any version unless the
node itself gets a direct invalidation. These tests also are using a "soft"
invalidation in a way that the storage does not encounter in any valid normal
workflow (in a normal workflow, a soft invalidation is always a result of a
deeper force dirty).

This updates the tests to add a dep to the node we are interested in and use
force-dirties on that dep to get the soft invalidations that we want.

Reviewed By: stepancheg

Differential Revision: D56154772

fbshipit-source-id: d20b6368df52f760254feb1e65bf64cae4e511c7