Fix VersionRanges::insert and remove use of SortedVectorSet
Summary:
We almost don't use any of the functionality provided by
SortedVectorSet and it just makes it harder for us to do interesting operations
on the VersionRanges.

This requires updating VersionRanges::insert, which was the one place that used
SortedVectorSet functionality. But that function is actually completely broken (
as can be seen from the test cases added in the previous diff). The updated
implementation is correct (or at least, "more correct").

Reviewed By: stepancheg

Differential Revision: D57318092

fbshipit-source-id: f8c93ba77e245cc9cbeddd995f24cfde7a1d1971