Optimize invalidate_rdeps
Summary:
The current invalidation requires O(E) allocation because it could add
a particular key to the queue many times (if it sees many edges to it before
actually processing it). This is a significant cost and even the extra HashSet
and insertions is small relative to what is saved.

Reviewed By: stepancheg

Differential Revision: D55547377

fbshipit-source-id: 1ca3bc8bd33676a4bc53e2d4b21f5c5931b7c752