Change introspection-serialized CellHistory
Summary:
I'm going to remove the use of CellHistory in modern dice. It's kinda
weirdly complex and has a bunch of bugs. Instead, we'll track a
`Vec<VersionNumber>` for force-dirtied versions (and stored forever) and a
VersionRanges indicating the valid versions.

This diff just updates introspection to have that form to make the next diffs
simpler.

Reviewed By: stepancheg

Differential Revision: D56154770

fbshipit-source-id: 22856b363b34b03eba47bfb162ce8862801d757a