Stop using CellHistory
Summary:
CellHistory is complex and (likely due to that complexity) there's
quite a few bugs in it and in its users (see the tests that no longer panic
after this change).

This changes us over to tracking history with two things:

VersionRanges: indicates the valid version ranges for a node
ForceDirtyHistory: a list of all versions at which the node is force-dirtied

The main thing this changes is how a node's valid versions gets expanded based
on the valid versions of its deps. In the new model, this is actually conceptually
really simple:

1. intersect the valid VersionRanges for all deps
2. intersect that with a VersionRange from the node's own ForceDirtyHistory to
ensure it doesn't cross a force-dirty.
3. done. union that with the VersionRanges that we know are valid for the node
(via other sources, for example, that we just computed it).

Not only is this conceptually simpler and easier to verify the correctness of, it
actually allows us to expand to more versions than the CellHistory based one
does. The cellhistory based one tries to do the same, except instead of using deps
full valid versionranges, it tries to get just the versionrange containing the
currently computed version (it does this incorrectly and sometimes includes versions
for which the dep isn't actually valid).

Reviewed By: stepancheg

Differential Revision: D56154786

fbshipit-source-id: b7b39206e9240b68748b6e1ab71e627168faabcd