Strip UNC path prefix on windows
Summary:
A user reported an issue with rust-analyzer on windows that we tracked down to these paths. In the future, we should probably make rust-analyzer understand them, but for now just strip the prefix since it shouldn't be relevant for any of our use cases.

For context: https://learn.microsoft.com/en-us/dotnet/standard/io/file-path-formats

Reviewed By: dtolnay

Differential Revision: D57642813

fbshipit-source-id: 8bd95b5d79b7b888017d11305c78a22b9debd76e