Simplify RemoteActionResult
Summary:
Adding `ActionCacheResult` to replace `Box<dyn RemoteActionResult>` for reporting results of action cache requests.

The motivation is so that I can move `action_result` out of `response` in the next diff without cloning, which I can't do with a boxed trait object.

This also means we only need 1 impl of `RemoteActionResult` instead of 3 that are identical except for `execution_kind`, which seems better.

Reviewed By: JakobDegen

Differential Revision: D57603035

fbshipit-source-id: 8daae5564949f5321d4f5eb3a8d5e4ea10f5021e