remote dep files: Upload dep file on action cache hit
Summary:
The fact that we only upload dep files on action execution is a problem because the builds that have permission to write to the action cache (on continuous/continuous_stable) are likely to get many cache hits from builds running elsewhere that trigger action cache writes from RE. We should upload dep files for those actions otherwise the number of dep file cache hits will be very limited.

We should also do this because we include the revision in remote dep file keys, so we should have dep files for cached actions on more than one revision.

Reviewed By: IanChilds

Differential Revision: D57651594

fbshipit-source-id: 6171b2481b9179ad771391dfd6283be3cf203a14