Enable library evolution if target-platform has it
Summary:
# Context
We want to provide an ability to compile XCFrameworks with Buck2 that contain Swift and can be consumed by Swift-based projects in Xcode.

To do this, we need to ape Apple's BUILD_LIBRARY_FOR_DISTRIBUTION Xcode setting. This enables swift-library-evolution and adds the module interface files to the resulting frameworks.

# This Diff
Adds an `_enable_library_evolution` attr to apple_library that is true if you add the configuration to your target platform. Then the arguments are passed into the compile.

Reviewed By: blackm00n

Differential Revision: D57619641

fbshipit-source-id: f6588790ccc4ade86a759bf4054af1e1a0238551