XCFrameworks will use swift library evolution
Summary:
# Context
We want to provide an ability to compile XCFrameworks with Buck2 that contain Swift and can be consumed by Swift-based projects in Xcode.
To do this, we need to ape Apple's BUILD_LIBRARY_FOR_DISTRIBUTION Xcode setting. This enables swift-library-evolution and adds the module interface files to the resulting frameworks.

# This Diff
apple_xcframework is powered by a transition for each of the underlying frameworks.
This diff adds the `:swift_library_evolution_enabled` to this transition so all frameworks will be built with `_enable_swift_library_evolution: True`.

Reviewed By: blackm00n

Differential Revision: D57620079

fbshipit-source-id: a67700f26a2a4ae382802318733f219619f095fe