Generate files for swift->swift xcframework consumption
Summary:
# Context
We want to provide an ability to compile XCFrameworks with Buck2 that contain Swift and can be consumed by Swift-based projects in Xcode.
To do this, we need to ape Apple's BUILD_LIBRARY_FOR_DISTRIBUTION Xcode setting. This enables swift-library-evolution and adds the module interface files to the resulting frameworks.

# This diff

When library evolution is enabled for the target platform, we are going to both pass this as a compiler option, and also request the generation of the files that support consuming the generated library from later versions of the swift compiler in Xcode.

Those files are the swiftinterface, the private.swiftinterface, as well as the swiftdoc, and will be encapsulated in this diff as an optional `SwiftLibraryForDistributionOutput` that is generated when library evolution is enabled.

We are not generating the abi.json file here. This is packaged up by Xcode along with the others, but I can't get swiftc to reliably generate it for all targets. It's sometimes just missing even when passing `-emit-abi-descriptor-path`. From searching online, a swift compiler author says this file is not needed for supporting library evolution:
https://forums.swift.org/t/xcode-14-creates-arm64-apple-ios-abi-json-by-default/65380/2

Differential Revision: D57620078

fbshipit-source-id: 0846810d7c486365e67b1b811c8f1dd3dc517f0c