Generate (single-arch) swift (XC)frameworks that can be used from swift
Summary:
# Context
We want to provide an ability to compile XCFrameworks with Buck2 that contain Swift and can be consumed by Swift-based projects in Xcode.
To do this, we need to ape Apple's BUILD_LIBRARY_FOR_DISTRIBUTION Xcode setting. This enables swift-library-evolution and adds the module interface files to the resulting frameworks.

# This diff
Adds and uses `_copy_swift_library_evolution_support()` to copy files into the frameworks for swift support.
Also adds a test
We currently only produce support for one arch there. Next diff adds support for files for both archs of a universal binary.

Differential Revision: D57434938

fbshipit-source-id: ece8dc63c8e8bbc88a69b4a4248e80187b680203