Add more config for modern dice rollout
Summary:
Right now dice is enabled with a single key, for non-repo-based
rollout (like gk-based) this poses the problem that if there's a bug in modern
dice we can't enable it for anyone while there are people using bad versions of
buck.

This change introduces a new key for non-repo-based
rollout:
modern_dice_min_version. We'll use
that to enable modern dice, and if there's a bad bug in buck we can bump the
configured modern_dice_min_version to disable it for any builds before the fix.

This also adds a separate RolloutPercentage configuration (modern_dice_rollout)
for when we get to the point that we want to rollout to those only covered by
in-repo configuration.

Reviewed By: JakobDegen

Differential Revision: D57753266

fbshipit-source-id: 443080fd50ae8a860327da94a610609e29326e3a