buck2 | Allow too-old target platform with a warning
Summary:
D57328419 added clamping logic to Catalyst builds, so we can build an
XCFramework that supports iOS12.0 for the iOS binaries. Before D57328419 the
Catalyst transition for these builds would try to build for Catalyst 12.0, which is lower than
the first version of Catalyst (13.1) and fail.

D57441943 made having a too-old platform an error, and builds started failing.

This diff switches this fail() to a warning() until we get some better logic
in place for this transition. (Better logic would also be able to handle macOS
platform constraints...).

Reviewed By: rmaz

Differential Revision: D57778990

fbshipit-source-id: bccba2a24cc83c7b6d663582803adc1ec7e42a4a