Create BXL for determining if files are type checked by Pyre PTT
Summary:
In the process of creating a mode header lint for Pyre, I was notified that querying Buck in a subprocess call for each file we're trying to check will be really slow (especially if we're doing multiple calls per file).

To speed this up, I turned it into a BXL where we can pass in all the files we need to query for at once and get back a JSON object mapping file to 'is type checked'. This should hopefully be a massive improvement to the earlier approach, and make it so that we can continue with setting up this lint.

Reviewed By: grievejia

Differential Revision: D57935901

fbshipit-source-id: 926c6d8d7dc136c681d530f1a89015d9024abdc3