clean up and fix test_executor configuration
Summary:
Follow up to D57952818, that diff didn't fix `*_with_compiled_buck2` tests since they set test_executor to an explicit tpx binary.

Tpx is used if test_executor is None or if BUCK2_TPX is set.

Refactoring so that `test_executor` doesn't need to be parsed from argv to make this easier.

Reviewed By: JakobDegen

Differential Revision: D58010186

fbshipit-source-id: 2255b15067c0b276ece259eff25d65f14fa8d3cb