Use proper preorder traversal for linkables
Summary:
We have [two types](https://fburl.com/code/6qafo8l5) of traversals for linkables: left-to-right preorder and topological.
Currently link groups are using right-to-left traversal that is distinct from not using link groups.

That causes problems for cases where link order matters. An example of it is linking asan runtime and libstdc++
By default libstdc++ goes as the last linkable. But in link groups we breaking it and we have runtime issues due to it.

I am fixing that distinction to better match non-link groups usecase. After this diff we use left-to-right traversal too.

Reviewed By: fgasperij

Differential Revision: D57905416

fbshipit-source-id: 74a4ddc795cc48445ed417743c66df10662411e0