Dump test executor heap using 'buck debug heap-dump'
Summary:
I noticed tpx is taking a lot of memory on my CI jobs, this stack makes buck2d dump its heap as part of buck debug heap-dump.

This diffs adds an arg to `buck debug heap-dump` to make it dump the test executor's heap. It only dumps the last executor's heap at the moment, which is good enough for me.

Reviewed By: krallin

Differential Revision: D58137504

fbshipit-source-id: 5ab2ffdbb149111ebf6d715a02f597e199178109