Add abstraction to handle paths behind symlinks
Summary:
This adds a new tuple-like record, `ArtifactExt`, to model paths
behind symlinks to other read-only filesystems.   This record
holds both the path to the symlink and the sub-path behind it.
The former is modeled as an `Artifact` which buck2 can track
and the latter is a string to append onto the former when referencing
the path in command.

Reviewed By: igorsugak

Differential Revision: D57117351

fbshipit-source-id: ba3efbf0566327635b746161f343eb2b823e2171