Ensure antlir plugins use static_pic link style
Summary:
Antlir feature plugins are specially-crafted rust dylibs that are used in order to plug in different feature to antlir2. They depend on the following properties:

1. The output file format is a dylib artifact.
2. The rust code underneath the artifact is statically linked in.

These are both true today, as we have `force_rlib=True`. However, when `advanced_unstable_linking` rolls out, this will no longer be true. In order to maintain the expected behavior here, I am exposing `link_style` to rust_library targets and setting it to `static_pic` for antlir2.

Reviewed By: dtolnay

Differential Revision: D58169449

fbshipit-source-id: 52c00b921a88579de74f4a5e9ef229b893629edd