Support timeouts when interacting with worker tools
Summary: Test Runway uses a worker-based local executor to distribute tests across multiple CI machines. This local executor has a concept of timeouts, which we can't communicate to Tpx right now because worker tools themselves don't support timeouts. This diff allows the worker to communicate that its execution timed out. See the previous diff for the API. Making the response an enum will probably break too many things, which is why I opted for a "timed_out" field on the response. Kinda lame, but I think this is ultimately safer.

Reviewed By: krallin

Differential Revision: D58188554

fbshipit-source-id: 1d94549c758c15ecd861f838fe9a25662421911f