Disable implicit propagation of shared linkage for nodes in the NO_MATCH group
Summary:
We transitively propagate shared linkage for nodes in link groups. Without this functionality, it's possible to e.g. pull the shared version of a prebuilt library without pulling in the shared version of its dependencies, which can lead to runtime errors.

There are some cases where we may want to disable this propagation logic. For example, when we're processing a manually or implicitly added dependency that is not actually needed at runtime. This diff allows opting out those dependencies by adding them to the special `NO_MATCH` group.

Reviewed By: Nekitosss

Differential Revision: D57826086

fbshipit-source-id: 207485402004285e6df5cb0a3e4de3fe38613174