Un-suffix metadata-fast actions, rename metadata-full actions
Summary:
As of {D57738355}, check builds are no longer based on a metadata-full action.

We used to use the action name `rustc check` for check builds, and a suffixed name like `rustc check [shared]` or `rustc check [fast]` for other sorts of metadata.

This diff converts metadata-fast to get the short `rustc check` name in check builds, and uses a suffixed name like `rustc metadata [static]` for the hollow rlib actions in the 3 different link styles.

| before | after |
|---|---|
| `rustc check [fast]` | `rustc check` |
| `rustc check` | `rustc metadata [static]` |
| `rustc check [pic]` | `rustc metadata [pic]` |
| `rustc check [shared]` | `rustc metadata [shared]` |

Reviewed By: capickett

Differential Revision: D58318938

fbshipit-source-id: 6af1fe11031648a39c2d4176cdf4e1804c196a06