rewrite code in fbcode/buck2/prelude/python/python.bzl to be python compatible.
Summary:
Starlark's is supposed to be a subset of Python, some tools (like autodeps) operate based on this assumption.

However, there are a few exceptions in corners, like this one. Using `dict({}.items() + {}.items())` to concat dicts is still allowed in starlark, but not in python 3 anymore. so this snippet of code works with buck but [will choke autodeps](https://fb.workplace.com/groups/autodeps.users/permalink/2661430264025456/) if any diff introduce it to the autodeps' scope.

modify the code to be compatible with both Starlark and Python, hence autodeps can handle it correctly.

Reviewed By: scottcao

Differential Revision: D58383682

fbshipit-source-id: ab830beb7fe5a7282f5883baf2096db1353e45f5