Remove outdated comment in bytecode interpreter
Summary:
Contest: `run_block` was called from two places: when evaluating `def` and for each `for` loop. A while ago `for` loops bytecode were rewritten without recursion (D45034981; to deal with native stack overflow, and to be able to implement GC properly), so this function is called only to interpret `def`.

But the function should not be inlined anyway, because the function body is huge (every instruction implementation is inlined into it, and if we inline it further, we slow down compilation at best, and probably decrease generated code quality). But this is somewhat obvious.

Anyway, remove the comment.

Reviewed By: ezgicicek

Differential Revision: D58369814

fbshipit-source-id: 8cae35a7c851acbb642082568ae4de8526959626