add transition for apple_binary propagated_target_sdk_version
Summary: We need a way to correctly set the target_sdk_version for a cxx_toolchain based on the version specified for an application. This change adds a transition from apple_binary that will set a target-sdk-version constraint if the attribute is set. This will be used to select on the cxx_toolchain to provide the correct flags for deps.

Reviewed By: milend

Differential Revision: D57626851

fbshipit-source-id: 56b08dd010fec4e0e69ea4d91d9b7643424d0652