Faster macro arg parser
Summary:
Most `attrs.arg` do not contain any macros or escapes. Use `memchr` to quickly rule them out.

This changes is correct because this is state transition from `Searching`:

https://www.internalfb.com/code/fbsource/[e57ffa9e3cb89ec48f15f44af421ef04e41e5ad0]/fbcode/buck2/app/buck2_node/src/attrs/attr_type/arg/parser.rs?lines=359%2C366-367

So we only search for backslash and dollar, which is what we do with `memchr`.

Result of calling `memchr` can be passed to `read_literal_opt_slow`, this is implemented in D58430978, but need to benchmark that change properly first.

Reviewed By: ezgicicek

Differential Revision: D58427524

fbshipit-source-id: d50a97bb8e42babc9a08b6ff121594f225160b76