EvalCallbacksMode
Summary:
`EvalCallbacksEnabled` only work for bc profiler or before stmt. Not both (because we don't allow running more than one profiler at the same time). And not none, because it would be pointless.

However, this diff reveals a bug: we attempt to initialize `EvalCallbacksEnabled` with no callbacks enabled. So temporary `EvalCallbacksMode::Disabled` is added.

Reviewed By: ezgicicek

Differential Revision: D58371988

fbshipit-source-id: 0a6d7da080f4d946ed28bd8efdfad888e4745151