Deny on Check Lints
Summary:
Introduces a new level of linter `deny_on_check_lints` which acts accordingly:

1. When running check builds (i.e. `[check]`, `[diag.json]`, `[clippy.json]` etc subtargets), treat the `deny_on_check_lints` lints the same as `deny_lints`
1. Otherwise, treat `deny_on_check_lints` the same as `warn_lints`.

This addresses a feedback that has come up a number of times for rust, in the context of deny lints and `deny(warning)`:
*  https://fb.workplace.com/groups/rust.language/posts/26061694763452459
* https://fb.workplace.com/groups/rust.language/posts/8129406413774569/?comment_id=8129532770428600&reply_comment_id=8131488860232991
* https://fb.workplace.com/groups/buck2users/posts/3606677236255298/

The benefit here is that we can now introduce a set of lints that will present an error at diff / check time, but are not serious enough issues to warrant preventing code compilation in general. You may now introduce unused must_use code in your local dev loop, as long as that unused must_use is addressed before the diff lands.

Reviewed By: zertosh

Differential Revision: D58318701

fbshipit-source-id: 2cd4bc9ca1a5caeec5320bb398d246d01e8f6162