Locate the temporary binary in a __binaries__ dir
Summary:
When Xcode runs a test, it picks up the binary in the output directory instead of in the xctest, even though the environment suggest it should use the binary in the xctest.

To work around this issue, we can create this binary in a `__binaries__` dir, which Xcode does not use.

Reviewed By: blackm00n

Differential Revision: D58491003

fbshipit-source-id: 27f2aeef252582f895758d6b9a625632cceba0d2