Use ptr_eq for types that do not override equals
Summary:
Expression

```
x = C
```

where `C` is constant can be compiled to `EqPtr` instruction when equality is pointer equality. Before this diff we did it for `bool`, `NoneType` and short strings, now we also do it for arbitrary types that do not override `equals`.

Note we always do `ptr_eq` before calling `equals`:

https://www.internalfb.com/code/fbsource/[6a92a6b3965bfd83b1f2858738492e71485ff6c9]/fbcode/buck2/starlark-rust/starlark/src/values/layout/value.rs?lines=795-798

Reviewed By: blackm00n

Differential Revision: D58060437

fbshipit-source-id: 33c7ebaf2d7c1ad2fa4585d86af972a9cfce11e2