Frameworks that appeared with Xcode 16b1 SDKs
Summary:
Some new frameworks that appeared in Xcode 16b1 for macOS and iOS.

I used the extracted SDK deps to find the new frameworks in the latest Xcode:

```
$> diff -urN \
    (jq '[.clang[].name, .swift[].name] | unique' xplat/toolchains/apple/sdk_deps/xcode_15.4.0_15f31d-clang17/iphoneos_arm64.json | psub) \
    (jq '[.clang[].name, .swift[].name] | unique' xplat/toolchains/apple/sdk_deps/xcode_16.0.0_16a5171c/iphoneos_arm64.json | psub)
```

The results are in P1417165413 (iOS) and P1417167783 (macOS).

I left out some frameworks that do not have public documentation. We can
probably guess the versions, but since they should not be used, I think it
should be OK to fail when trying to use them.

I also left out the new Testing framework. XCTest does not seem to be in the
list, so it is probably handled in some other way.

Docs:

- https://developer.apple.com/documentation/AccessorySetupKit
- https://developer.apple.com/documentation/ContactProvider
- https://developer.apple.com/documentation/CoreHID
- https://developer.apple.com/documentation/fskit/
- https://developer.apple.com/documentation/LockedCameraCapture
- https://developer.apple.com/documentation/mediaextension/
- https://developer.apple.com/documentation/mididriverkit
- https://developer.apple.com/documentation/proximityreaderstub/
- https://developer.apple.com/documentation/stickerkit/
- https://developer.apple.com/documentation/Synchronization
- https://developer.apple.com/documentation/TabletopKit

Reviewed By: rmaz

Differential Revision: D58623936

fbshipit-source-id: 96aeb24ff381119232e06741116163e035aabd26