Added instrumentation attrs to cxx_python_extension
Summary:
Add `coverage_instrumentation_compiler_flags` attribute to `cxx_python_extension`. This attribute will be set always and will contain the
compilation flags that are required to instrument sources. Prelude code will only used them when coverage is enabled for a target due to a header being selected for coverage.

Reviewed By: Nekitosss

Differential Revision: D57494354

fbshipit-source-id: 44abe0d648dc4db72da7be703df64a1adff79634