Add compression
Summary:
Doesn't change the performance, but does significantly reduce the size of the output - using the BTD command line the values are:

* Uncompressed - 7539Mb
* Gzip - 297Mb
* Zstd - 258Mb

Given we have results showing that Zstd speeds things up for reading and uses less bandwidth, might as well generate it that way.

Not that compression requires --output. I would have liked to generate compressed data over stdout, but the gRPC protocol we use makes the output a String, which can't take arbitrary bytes.

Reviewed By: stepancheg

Differential Revision: D58791465

fbshipit-source-id: 344cf03d78d9f8e38ea6fc9ea76b7766c23bbeb1