github: build aarch64-linux binaries with 16k-page jemalloc
jemalloc builds in the system page size as part of its (internal) ABI at compile
time, which means that if you build with a smaller page size than what your
system allows, it doesn't work.

In particular, this impacts Linux on Apple Silicon, which doesn't natively
support 4k pages which are the default. Set it to 16k so that the prebuilt
binaries work OOTB.

Signed-off-by: Austin Seipp <aseipp@pobox.com>