buck2 | Allow explicit opt in to Swift Library Evolution
Summary:
We need to enable library evolution in the context of XCFrameworks.

1) It allows for forward compatibility
2) Without it you can't generate the swiftinterface files.

However, the current approach automatically enables evolution for entire XCFramework builds. This is undesirable.

1) It frequently introduces thousands of compiler errors from dependencies that are incompatible with Swift Library Evolution
2) Those dependencies are not public to the finished library / framework. The top-level apple_library is able to expose itself as a single module with swiftinterface files, so only this target needs to be compiled with evolution.
3) It might have performance implications to enable it unnecessarily in parts of the compile.

For this reason, this diff introduces a new flag enable_library_evolution.

When this is enabled, a given apple_library target will be compiled with library evolution enabled.

It also removes the addition of the enable_library_evolution constraint from the XCFramework transition, to prevent the entire graph being built with evolution enabled.

Reviewed By: milend

Differential Revision: D58467395

fbshipit-source-id: d2c613ae0319317f385aca0134809b5327d7421b