Unify (somewhat) recursive and non-recursive analysis
Summary:
This diff changes how analysis profiling works internally.

Before we had:
- for recursive profiling, we stored one profiling mode globally, invoked profiler normally, and then collected profiling results
- for non-recursive profiling, we just invoked analysis without dice

Now we store on DICE patterns we should profile. Analysis, when runs, passes configured target to a new DICE key to find profiling mode for this key.

We resolve target patterns on demand (because when we initialize profiler mode, we don't have access to DICE and patten resolver). That's why we store working dir on DICE ~~in D58631198~~ in `StarlarkProfilerConfiguration`.

This is not very useful change for now, but it is a template to implement recursive loading profiling.

Reviewed By: cjhopman

Differential Revision: D58631203

fbshipit-source-id: 8de6af961093d7a28400bfa9e3a77fefada16eee