Add alias support
Summary:
Add alias support for CLI modifiers.
This diff adds an `ALIAS` struct that contains a list of aliases that can be used for CLI modifiers.
When resolving CLI modifiers, cfg constructor will first attempt to resolve any alias. It detects whether a string should be an alias or target based on the existence of `:` in the string.
Added aliases for OS and CPU constraints in this diff.

Reviewed By: samkevich

Differential Revision: D58368110

fbshipit-source-id: 50de7a6baef34132dbf6d2f51bbb5b74c4385055