Correctly pass linked against dylibs to final link
Summary:
Distributed thin-lto linker invocations previously failed to include linked against dylibs. I can only assume that FBCode links against libraries with search path based flags (ie. `-lFoo`), whereas for FBShared at least, we pass a path to the dylib directly to the linker.
This was previously reverted for S425211. To make sure this change is not affecting non-ios dthin-lto traffic, of which there is none, check that the linker type is darwin before taking the new codepath.

Reviewed By: rmaz

Differential Revision: D58823629

fbshipit-source-id: 9d1556a0d54a91be8f555b0728d176ec57bec9b8