Introduce dist_thin_lto_codegen_flags field to cxx_toolchain
Summary:
This new field stores flags that will be used by opt + codegen actions in distributed
thin-lto. This will include some of the linker flags, but not all, and some will be transformed
the linker is not involved in such an action.

We add an assert, and a more reasonable default to make it 100% clear that the new codepaths
are only exercised by ios dthin-lto builds, of which there are none at the moment.

Reviewed By: rmaz

Differential Revision: D58824005

fbshipit-source-id: 4bda804e42f6911f2cf5d3b9d1e09305a9acfbfe