Add `cell` attribute for StarlarkFileNode
Summary:
# Context
StarlarkFileNode originally only have `path` attribute api for StarlarkFileNode. It would make users to parse the cell manually.

# This diff
This adds cell attribute to `StarlarkFileNode` it.
In the task, it also mentioned adding getting`package` to `StarlarkFileNode`.
I think `package` for a `StarlarkFileNode` (a path to file) does not make sense here.

Reviewed By: cjhopman

Differential Revision: D58963491

fbshipit-source-id: c735e4e55f684d8b4502f5479076b4f952b64079