Add new ProgressHeader
Summary:
This gives a more detailed view of build progress. The most detailed
view requires some 130 or so columns, so it has some complexity to support
rendering different levels of detail. See the test for examples of output at
different widths.

This doesn't switch to the new header. Instead it's only enabled when
pressing 'x' during a build.

Reviewed By: stepancheg

Differential Revision: D58715675

fbshipit-source-id: c4fc8fb50d0682e9fd05c5d1e94362e6ae83f8e4