Extract TasksHeader
Summary:
This extracts a component that handles just the line above the tasks.
I'm going to explore a new version of this.

Right now the tests are mostly copies of the TimedList tests (that were
previously testing this code) but edited down to just test the new header. They
are unnecessarily complex imo and the next diffs clean them up.

Reviewed By: JakobDegen

Differential Revision: D58715677

fbshipit-source-id: e466d67cde8af0b8579eb3f0c5350d8172ae61c9