external cells: Fix cancellation bug
Summary:
We have a flakey test: https://www.internalfb.com/intern/test/844425079615050 .

The root cause of this bug is that when multiple keys are waiting on a git repo download to the same directory, they synchronize and the second blocks on completion of the first one. However, if the first key's computation then gets cancelled it looks like it finished and so the filesystem is left in an invalid state. Prevent that by not allowing this work to be cancelled.

Reviewed By: stepancheg

Differential Revision: D58921239

fbshipit-source-id: 7ff902067c9ff7f345f8f8af3ef907203d7a3a4e