Add ability to customize names used for constraints in naming function
Summary:
Certain constraints don't have good constraint names to use in configuration name directly. For example, the internal cuda constraint doesn't have the word cuda in the name. Adding an optional transform to each constraint setting that can be used for naming so that the name can be customized for each constraint value. The transform is a function that accepts a constraint value target label and returns a string.

Added clang constraint + transform to the naming function.

Reviewed By: get9

Differential Revision: D52304536

fbshipit-source-id: e92abba3629ece68e7895f043f5a9ff21d55e7ff