Check system_memory from recorded event not from host machine
Summary:
Currently, `buck2 replay` wouldn't display memory warnings accurately because
-  we were getting system memory from the host machine but not from the actual machine we built on.
- we didn't pass the buckconfig to replay to pick threshold from.

With D58812046,  we store the actual system memory as an instant event. This diff ensures that we use that memory (instead of host's memory) to display warnings via using the host's buckconfig. This is not perfect as we don't store the threshold from the user's machine. This could be improved later but I don't think buckconfig should change that much tbh so not prioritizing it now.

Reviewed By: stepancheg

Differential Revision: D58812023

fbshipit-source-id: d53ebe7d8cc5e695a4a0409a460e3847b71862d4