keep going: Remove support for `Unordered`
Summary: The `Ordered` version actually has a bug, which I'll fix later in this stack. However, fixing that bug while maintaining support for `Unordered` is a bit verbose and painful, so let's just support only the `Ordered` variant. The perf difference between the two at the end of this stack is only a few allocations.

Reviewed By: cjhopman

Differential Revision: D58964231

fbshipit-source-id: 80504c877057db546e03c370c47265796e584674