external cells: Add cell name in bundled cell origin
Summary:
`ExternalCellOrigin::Bundled` is an incomplete description of where the cell comes from. We also need to know the cell name to know which bundled cell we want. So let's store that in the enum.

This allows for a little bit of simplification next diff. I was hoping for more, but unfortunately didn't pan out

Reviewed By: stepancheg

Differential Revision: D58921236

fbshipit-source-id: 395ebda8b996ed9bb1d7e9f74716fcb668219055