github: build aarch64-linux binaries with 16k-page jemalloc (#693)
Summary:
jemalloc builds in the system page size as part of its (internal) ABI at compile time, which means that if you build with a smaller page size than what your system allows, it doesn't work.

In particular, this impacts Linux on Apple Silicon, which doesn't natively support 4k pages which are the default. Set it to 16k so that the prebuilt binaries work OOTB. **EDIT**: Apparently the Raspberry Pi 5 ships with a 16k page kernel by default, as well.

Pull Request resolved: https://github.com/facebook/buck2/pull/693

Reviewed By: samkevich

Differential Revision: D59100007

Pulled By: JakobDegen

fbshipit-source-id: cf462b2e89c72c949a6a06c1187df98210a5e3ea