Add UnpackedBuckEvent variants for unrecognized events
Summary:
Adding a new instant, span_start, or span_end type causes old versions
of buck to return an error from unpack_event when encountering that event. This
breaks replay and other tools when using a buck older than the new event but looking
at a newer build.

Instead, add new variants to represent that. Most uses of unpack_event are
already using it just to identify specific events and are ignoring the rest, and the
other use is easy to update.

Reviewed By: iguridi

Differential Revision: D59238048

fbshipit-source-id: 663a2bf0bf2b983f4459fae6578a293ebff8149b